api-tests/requirements.txt
pytest
requests
pydantic
faker
allure-pytest
pytest-mock